Enable board overlay format in board configuration
The camera is currently configured to YCrCb_420_SP, which is an Adreno
implementation of NV21. This format is causing corruption in the
images, but changing the format within the HAL leads to further
corruption, so instead it will be forced on the board configuration.
The frameworks being used is configured to allow switching to
YCbCr_422_I, which is Adreno YUY2, but currently the HAL and overlay do
not support this.